Update patch set 19
Patch Set 19:

I ran some tests over the long break on PS6 (spin lock) and PS11 (mutex)

On the Nexus 7 2013 I saw about a 1% increase raw decode overall on PS6 over PS11 for 720p and 1080p when using 2 threads. I saw about a 12% increase raw decode overall on PS6 over PS11 for 720p and 1080p when using 4 threads.

I saw about the same numbers on a Nexus 4.

I also possibly saw a deadlock on the PS6, but hard to know for sure. If it was, it was 1 in over 1000 playbacks.

Patch-set: 19
Label: Code-Review=0
Label: Commit-Queue=0
Label: Verified=0