Update patch set 2
Patch Set 2: Abandoned

Agreed, this is a bit too hacky.   If any one has a simple suggestion for a fix..   I'm happy to try.  

examples.mk generates a reasonable rule for vpxenc but without a dependency that generates libs. 

and 

the Makefile at root doesn't know to connect a make target like vpxenc to examples.mk...     

That was the reason for me simple ugly hack.

Patch-set: 2