thumb: Clarify a comment about the conversions that are done
The comment made it look like the condition code was dropped from
the extra add instruction, while it actually was handled properly.

Thus, the comment was misleading while the code itself did the right
thing.

Also clarify the comment indicating that we use the full three-operand
form of the add instruction.

Change-Id: I2c1ac6ac4fedf262d104ea30a6c005febc74de9c