Update patch set 1
Patch Set 1:

as I said yesterday, we can make the modification on import, no need to muddy the history here. moving forward now that we have a c++ compiler requirement for the examples we can update to scale.cc and make the change at that time.

Patch-set: 1
Label: Code-Review=0
Label: Commit-Queue=0