Update patch set 1
Patch Set 1:

+1 from Marco on vp9_spatial_svc_encoder. If that's adequate: Fine with me. 

I'm of the opinion that vp9_svc_encoder is acceptable as long as the header comment notes that it's spatial svc only, but I don't want to argue about this for a week.

So, @minghai: Is vp9_svc_encoder with a source file header comment OK, or do you prefer vp9_spatial_svc_encoder?

Patch-set: 1