Update patch set 3
Patch Set 3: (3 inline comments)

Generally speaking I like the idea of supporting coding a lower bit-depth source with a higher bit-depth codec in a more comprehensive way. However I am not sure if we should allow a higher bit-depth source to be coded at a lower bit-depth by downshifting the source. It seems that adds too much code complexity for a not-so-useful use case. Further it can be confusing for the user. So I would like to make a suggestion. Pick one of the following:
1. If you want to keep the downshifting option, at least add a warning with a prompt when input-bit-depth > codec-bit-depth (we do that in some cases ex. where user chooses --min-q and --max-q the same)
2. Remove the cases of downshifting the source and do not allow input bit-depth to be > codec bit-depth, as before. High to high conversion is good to have, but only as long as input bit-depth <= codec bit-depth.

Patch-set: 3
Label: Code-Review=0
Label: Commit-Queue=0
Label: Verified=0