Added config-enable-hw-high
In the inverse transform code, overflow results in
implementation defined behaviour.

Hardware implementations are expected to wrap on
overflow in order to use the smallest data path
possible.  The new config option makes the C reference code
use wrap on overflow.

Encoders are recommended to not generate bitstreams that
result in overflow in the inverse transform.

Change-Id: I73836ccfb925b4dbfb29c257a3784fa7733db877