Update patch set 1
Patch Set 1:

Also, one question: Should I be concerned that x86inc is being turned on for armX-darwin builds?

It's not just the arm64 target. It's all of them, and it seems really weird to me... although I'll admit to being almost completely ignorant of what's going on in there.

Patch-set: 1