Modify initial value for avg_frame_qindex
In high bitdepth mode there is an extended quantizer range.
This means that it takes longer for the avg_frame_qindex
to ramp up and this results in coding loss for short
low bitrate sequences.

This patch adds a boost in high bitdepth mode to compensate
for this effect.

This helps with the bowing_cif.y4m sequence at 50kbps.

Change-Id: Ie1575d88e8de4f0297cf86da50eb36dfc5442c70