Update patch set 1
Patch Set 1:

strictly speaking there's no prohibition for common reaching into decoder, but logically this may make more sense; in any case it was likely to move here when the encoder started using it.
note the requirement is actually the pthread emulation layer, not the worker's themselves in this case. That could be split into a separate module moving forward (and reused in vp8).

Patch-set: 1
Label: Code-Review=0
Label: Commit-Queue=0