Changes quantization tables to improve quality
Changes quantization tables so that the scale factor diff against
8-bt version is 1 at the low q end and increases to 4 (16) at the
high q end for 10 (12) bit versions, using a logarithmic like
function.

This change improves the quality gap aginst the no-high-quant version,
for both 10-bit and 12-bit. 10-bit is virtually similar in performance
to the no-high-quant version, while the gap for 12-bit is reduced to
the 0.3% range.

Change-Id: I08cc329a77aeb19cc232f8e4458f188c6ae65363