Update patch set 3
Patch Set 3:

> I've updated to add to CFLAGS for SSE2 or later on OS/2. But should I add all CFLAGS and CXXFLAGS on OS/2 ?

this is ok if it's working for you and won't be a problem with older compilers. if needed we could shift this to configure and broaden the application of the flag. I wonder if this is  the root cause or there's some other underlying alignment issue this is masking

Patch-set: 3
Reviewer: James Zern <1002219@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0
Label: Verified=0