Update patch set 2
Patch Set 2: (2 inline comments)

I think we have to think through the different scenarios that are possible now that the color space and/or subsampling can change on a non-keyframe.

How do we handle prediction in the case of such a change, for example.

Patch-set: 2
Label: Code-Review=0
Label: Commit-Queue=0
Label: Verified=0