Update patch set 1
Patch Set 1:

I can't help but think this is the wrong place for these counts and estimates.   If you do this per tile we'll need an extra sum..   

I believe the counts should be done as part of the rd loop and zeroed at the beginning of the loop.   As of now these mode counts are zeroed 2 times even if we don't recode and for realtime we never recode.   

Also as it stands in order for parallel to work if these are done in separate threads,  you'll need somethign to go and add all the counts together from each threads count set,  which would be totally unneccessary if it was just done in the pick_rd code.     

I will examine..

Patch-set: 1
Reviewer: James Bankoski <1000963@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=0
Label: Commit-Queue=0
Label: Verified=0