Update patch set 1
Patch Set 1:

There are two issues here, the boost and the availability of a older reference frame that may be used for repair even if not boosted. I suspect quite a lot of the "recovery or background" may  be possible even with no boost or a minimal boost, especially if the background in that older frame has been cleaned up by earlier application of cyclic refresh.

Patch-set: 1