Update patch set 2
Patch Set 2:

One issue is that there are up to 8% performance loss in a few stdhd sequences at the low end rate when using 4-tile coding as compared to single tile.

Ran experiments to verify that this is not related to this CL. Trying to figure out the reason that caused the difference between single- and 4-tile.

Patch-set: 2