Update patch set 8
Patch Set 8: Looks good to me, approved

Re use of the overshoot_pct. I guess this is OK for now but it means we have a very different meaning for this command line parameter for 2 pass and 1 pass CBR. In two pass it specifies a tolerance to overshoot before we take drastic corrective measures, here it is being used to indicate the initial boost ratio.

I accept though that this is probably a temporary / experimental situation.

Patch-set: 8
Label: Code-Review=+2
Label: Commit-Queue=0
Label: Verified=0