Update patch set 9
Patch Set 9:

A couple of comments;
I guess you will need to rebase since there is now the 64x64 transform also in the nextgen branch. It seems your code is based on a version prior to that.
There is another experiment on using alternate transforms for inter (EXT_TX) in the playground branch, which I was about to put into the nextgen branch. No transform is just a different kind of  transform (identity), and you may be able to reuse a lot of the code from that experiment into the playground branch, including rd decisions.

Patch-set: 9