Update patch set 1
Patch Set 1:

> But the problems are:

> 1. Why this error happens with my patch now? Before my

> patch, cm is also accessed in the handler? And this error does not happen in

> frame parallel branch.

Likely it's optimization related only triggered if the compiler is considering
reworking access. This was a latent issue from the beginning, again: see the
code in webp jpegdec/pngdec for reference.

> 2. Why the error is only about cm, but not pbi?

It's only for local variables (not function parameters) accessed between calls
to setjmp() as described in the man page.

Patch-set: 1