Update patch set 3
Patch Set 3: Do not submit

As it stands this will break the visual quality improvements from lowering the cost of newmv where there is no non  0 nearest or near. (I44a29e20f11f7ab10f8c93ffbdc50183d9801524).

If the best error for 0,0 is low (as it often is in sky scenes for example) we will never get the chance to check the new mv and establish a motion field. It should eb possible to limit the speed implications by only resetting the mode check threshold to 0 in cases where we have not found a non zero nearest or near.

Patch-set: 3
Label: Code-Review=-2
Label: Commit-Queue=0
Label: Verified=0