Update patch set 7
Patch Set 7:

Should we tie it exactly to the maxsection-pct as this is also used to relax maxQ from the active value out to the user specified maximum. It may be the right thing to do but I would not want to see us forced to set a very high maxsection-pct just to avoid scaling too often.

Patch-set: 7
Label: Code-Review=0
Label: Commit-Queue=0
Label: Verified=0