Update patch set 3
Patch Set 3: Code-Review+2

I'm going to land this as is. On a few stable testing systems I see a slow-down when manually unrolling all instances x4, x8 is mixed. There's more work to be done here and if someone wants to extract the individual functions, unroll and time them later and tell me I'm wrong, that's OK.

Patch-set: 3
Label: Code-Review=+2