Update patch set 6
Patch Set 6:

(1 comment)

on various devices patchset 6 is slower than patchset 4; though it is a good cleanup, this points to the function needing a larger overhaul (to reduce register spill). for now restore patchset 4 and add a TODO to remove the transpose along with a note about the performance drop.

Patch-set: 6
Label: Code-Review=0