Update patch set 5
Patch Set 5:

(1 comment)

wait for Yaowu's [1] change to land and add a quick note to the CHANGELOG.
this is unfortunate that we are moving the ABI check for the existing code only for people who would attempt to build an old example or code against a newer released lib. with the current state I don't see a way to avoid this while still indicating that this change has made an update.

[1] https://chromium-review.googlesource.com/#/c/274401

Patch-set: 5
Label: Code-Review=0