Update patch set 1
Patch Set 1:

The C functions were declared in vpx_dsp/inv_txfm.h.

They are intended to be non-static. We have 1-D DCT and 1-D ADST functions separately. In the vp9_idct.c functions, there are wrapper functions that call a combination of DCT and ADST depending on the transform type.

Patch-set: 1