Check CONFIG_VP<N> instead of encoder and decoder
If either the encoder or the decoder is enabled, CONFIG_VP<N> will be
set. This simplifies the conditional and passes the chromium update
script when CONFIG_ values are passed in with 'yes' and 'no' values.

This was failing because it was checking against empty strings but
they are set to 'no'

Change-Id: I02ecd557210088ba1458cd0e89eead5666f6597a