Update patch set 5
Patch Set 5:

I'll look into vp8, I think it should just set display_width/height values to width/height. As for removing it in vp10, that's your decision. I can see some utility in it in theory, but right now it's wasted because it's never used. If you don't think that's worth it (or it's a container thing), that's OK with me and I can remove it in vp10, yes. If we expose it in ffmpeg, it suggests we think it has utility and we may want to keep it...

Patch-set: 5
Label: Code-Review=0