Update patch set 3
Patch Set 3:

Correct me if I am wrong.

I would recommend not to use #if statement in the master branch. This can reduce the readability.

Anything we merge into the master branch should be on by default.

How about to get the code in with an encoder parameter control to turn it on only for screen content?

Patch-set: 3