Update patch set 1
Patch Set 1:

Not sure if this should be done, or if we should just change the automated stuff to use --extra-configure-args. The only thing I can really say that makes this better than the alternative is that --test-link is dead obvious to someone that doesn't know why we're passing --enable-examples.

Patch-set: 1
Label: Code-Review=0