Update patch set 1
Patch Set 1:

Yeah Bruce spotted it. I've been aware of it for a while and may have looked into the callers to see that they aren't sending 0s? But that's why I added Alex.

The implementation in the quantize does check for zero because it is a possibility there.

Patch-set: 1
Label: Code-Review=0