Update patch set 6
Patch Set 6:

So, this is extremely frustrating. The bug pointed out by asan is indeed reproducible, but has nothing to do with my experiment. I can trace it back (using bisect on test_libvpx --gtest_filter='*VP10/VPxEncoderThreadTest*' with --enable-misc_fixes) to 7238492235237457997cfa8a9713ea36523174cb, but that commit seems to have nothing to do with the actual bug. The bug is already present before that, and I'm going to have a real fun time bisecting it, I guess... The real issue here is that the ethread unit test (or any test, for that matter) does not use the same mechanism that vpxdec --test-decode=fatal uses. That is, the code that currently lives in vpxdec should be refactored and placed somewhere so all unit tests (or at least the ethread ones) use it also.

Patch-set: 6
Label: Code-Review=0