Update patch set 1
Patch Set 1:

> What's nextgenv2?

nextgenv2 is a branch which is supposed to be the staging ground for all new tools. This is where we will move all tools (either from nextgen or new) with experimental flags. Once tested/verified and is deemed fast enough, we will merge them without the flag into master. There will be periodic merges made from master to nextgenv2 so that nextgenv2 does not go out of sync too much in the non-vp10 directories.
So ideally misc_fixes (and anything else that is a bitstream change) should have been in nextgenv2 first, before we merged with master and removed the flag. There was a confusion initially regarding this process - sorry for that. Currently nextgenv2 is being fixed for some test failures because some of the jenkins tests were not on, but that should be taken care of in a few days.

Patch-set: 1
Reviewer: Debargha Mukherjee <1001184@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Verified=0