Update patch set 2
Patch Set 2:

Yunqing - this code was added to work around textrel: 
https://chromium-review.googlesource.com/#/c/67765/

Sergey found the issue in the linked bug.

The fix looks correct to me:
GET_GOT is indeed a non-op for win32. It's only defined for elf32 and macho32. Using 'add esp, 4' instead of RESTORE_GOT looks like a mistake.

This leaves the GOT in place for the GLOBAL[] and cleans it up correctly.

Just wanted to add you to check if GET_GOT *sholud* be defined for win32?

Patch-set: 2
Label: Code-Review=0