Update patch set 2
Patch Set 2:

My point is that TSan is highly unlikely to misunderstand the synchronization here (we do intercept sem_post() and sem_wait()), so the warnings may denote actual bugs.

Let us take a look at lines 860-866 of vp8/encoder/encodeframe.c.
In the case (mb_row == cm->mb_rows - 1), i.e. we're executing the last iteration of the mb_row loop, the following happens:
 - on line 862 sem_post() increments the value of semaphore |cpi->h_event_end_encoding|, making it greater than zero;
 - because this is the last iteration of the loop, the execution proceeds to line 866;
 - on line 866 sem_wait() decrements the value of |cpi->h_event_end_encoding|. Because it was greater than zero, and nobody else decremented it (we're on the main thread), the value of the semaphore value becomes >=0, and sem_wait() immediately returns.

The above contradicts with the comment on line 866, which says "/* wait for other threads to finish */" - sem_wait() does not wait for other threads in this case.

Patch-set: 2