Update patch set 2
Patch Set 2:

I think the instability is caused by blk_skip.

As far as I can see, blk_skip is only set for plane==0 within VAR_TX blocks.
However, in SUPERTX blocks, blk_skip is set for all planes (if the supertx block is skipped).

This means that if we use a skipped SUPERTX followed by a VAR_TX, some of the VAR_TX UV blocks will be skipped.

This ends up causing multi-thread instability because the contents of blk_skip persist across tiles.  I would also expect this bug to cause loss of compression quality as random chroma blocks are being skipped.

The fix is not quite clear to me as I don't fully understand the purpose of skip_recode and blk_skip: perhaps the answer is to set blk_skip for planes 1 and 2 to be equal to skip for non-SUPERTX blocks?  Or perhaps the answer is to simply remove blk_skip for planes 1 and 2?

Patch-set: 2
Reviewer: Peter de Rivaz <1138583@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=0