Update patch set 1
Patch Set 1: Code-Review+2

Was this a maybe-uninitialized with some toolchain? not sure how this one slipped by if the per-commit would have caught it. note we don't fail multi-platform builds for warnings due to legacy vp8 -- we could put a cap on that (or fix them) though

Patch-set: 1
Label: Code-Review=+2