Update patch set 3
Patch Set 3: Code-Review-2

The basis of the problem is that tran_[low|high]_t depends on what configuration the code is built in:
https://chromium.googlesource.com/webm/libvpx/+/master/vpx_dsp/vpx_dsp_common.h#26

This means that the intermediary buffers may be different sizes. In this change:
https://chromium-review.googlesource.com/324310
the buffer-handling code is ifdef'd to account for the different possibilities.

It may be the case that this code does not appear to be affected: perhaps the buffers are only passed between optimized versions of the function and through this happy accident the data is correctly handled.

This would need to be fixed/verified by applying a fix similar to the one above where the different size buffers are accounted for or by finding a way to use different intermediate buffers for the 8 bit case.

Patch-set: 3
Reviewer: Johann Koenig <1117029@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=-2